Template: vite-cf-DO-runner
Base template for Build a clone of the Kahoot quiz game using Durable Objects. It should take a list of mutliple-choice questions with 2-4 answers. Only one answer can be the correct one. People join a game by scanning a QR code. Initially user enter their name and join the game. When the admin starts the quiz, buttons with the colors of the answer displayed in a random order on the screen of the players mobile device. It should use WebSockets to show the current answer options (as colored buttons) and track the answer. The players are scored by correct answers and how quickly they answered. The question and answers will only be displayed on the main screen by the person running the game. Users have 20 seconds to answer the questions.